+++ utils/general.py
@@ -7,7 +7,7 @@ def separate_symbol_and_trade_against(symbol):
     if symbol.endswith("BTC"):
         symbol_only = symbol[:-3]
         symbol_stable = symbol[-3:]
-    elif symbol.endswith(("BUSD","USDT")):    
+    elif symbol.endswith(("BUSD","USDT","USDC")):    
         symbol_only = symbol[:-4]
         symbol_stable = symbol[-4:]